Django_Blog/src/requirements.txt
adodbapi==2.6.2.0
asgiref==3.2.10
astroid==2.4.2
attr==0.3.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bootstrap4==0.1.0
bqplot==0.11.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
decouple==0.0.7
defusedxml==0.6.0
Django==3.0.7
django-bootstrap4==2.1.1
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-tempus-dominus==5.1.2.13
django-tinymce4-lite==1.8.0
docopt==0.6.2
entrypoints==0.3
idna==2.10
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
jsmin==2.2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
nbinteract==0.2.5
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tests==0.7
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
traittypes==0.2.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zmq==0.0.0
